+++ NettyDockerClient/src/main/java/client/json/configs/hostconfig/portbinding/PortBinding.java
@@ -1,4 +1,4 @@
-package client.json.configs.hostconfig;
+package client.json.configs.hostconfig.portbinding;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
@@ -7,6 +7,7 @@ import java.util.List;
 import java.util.Map;
 
 public class PortBinding {
+    //TODO 한번 더 고민해보기
     private Map<String, List<HostPort>> portMap = new HashMap<>();
 
     public void setPortMap(String exposedPort, List<HostPort> hostPorts) {
